+++ app/src/main/java/de/foodbro/app/ui/edit/RecipeEditViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.launch
 
 class RecipeEditViewModel @AssistedInject constructor(
     private val recipeRepository: RecipeRepository,
-    @Assisted val recipeId: Int
+    @Assisted val recipeId: Long
 ): ViewModel() {
 
     val recipeDetail = recipeRepository.getRecipe(recipeId)
@@ -19,13 +19,13 @@ class RecipeEditViewModel @AssistedInject constructor(
 
     @AssistedInject.Factory
     interface AssistedFactory {
-        fun create(recipeId: Int): RecipeEditViewModel
+        fun create(recipeId: Long): RecipeEditViewModel
     }
 
     companion object {
         fun provideFactory(
             assistedFactory: AssistedFactory,
-            recipeId: Int
+            recipeId: Long
         ): ViewModelProvider.Factory = object : ViewModelProvider.Factory {
             @Suppress("UNCHECKED_CAST")
             override fun <T : ViewModel?> create(modelClass: Class<T>): T {
